refactor: file structire, naming, standard paths
Addresses feedback from review on moved stores wait
and analysis to pinia pr.

File structure and naming now follows the reccomended
structure from Pinia, rather than the previous vuex model:
https://pinia.vuejs.org/cookbook/migration-vuex.html#Restructuring-Modules-to-Stores

Standardized paths to @/ rather than relative paths.
