+++ src/stores/analysis.ts
